+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.MyApplication">
         <activity
-            android:name=".ui.MainActivity"
+            android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.MyApplication">
@@ -31,7 +31,7 @@
             android:exported="false"/>
         <activity android:name=".ui.check.CheckNameActivity"
             android:exported="true"/>
-        <activity android:name=".ui.check.EditNameActivity"
+        <activity android:name=".ui.check.CheckResidentNumberActivity"
             android:exported="true"/>
     </application>
